% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regarima_spec.R
\name{add_outlier}
\alias{add_outlier}
\alias{add_ramp}
\title{Add Outliers/Ramps to Specification
Generic function to add outliers to a specification}
\usage{
add_outlier(x, type, date, name = sprintf("\%s (\%s)", type, date), coef = 0)

add_ramp(x, start, end, name = sprintf("rp.\%s - \%s", start, end), coef = 0)
}
\arguments{
\item{x}{the specification.}

\item{type, date}{type and date of the outliers. Possible \code{type} are:
\code{"AO"} = additive, \code{"LS"} = level shift, \code{"TC"} = transitory change and
\code{"SO"} = seasonal outlier.}

\item{name}{the name of the variable (to format print).}

\item{coef}{the coefficient. If equal to 0 the outliers/ramps are estimated.}

\item{start, end}{dates of the ramp regressor.}
}
\description{
Add Outliers/Ramps to Specification
Generic function to add outliers to a specification
}
